requirements.txt
# Core Framework  
fastapi==0.115.14  
uvicorn[standard]==0.35.0  

# Database  
SQLAlchemy==2.0.41  
PyMySQL==1.1.1  
sqlmodel==0.0.24  # Optional if you're using SQLModel  
psycopg2-binary==2.9.9  # For PostgreSQL support (needed in some deployments)

# Security & Auth  
passlib[bcrypt]==1.7.4  
python-jose==3.5.0  
email_validator==2.2.0  
bcrypt==3.2.2  
python-dotenv==1.1.1

# Additional dependencies
requests==2.32.3
annotated-types==0.7.0
anyio==4.9.0
cffi==1.17.1
click==8.2.1
colorama==0.4.6
cryptography>=4.0.0
dnspython==2.7.0
ecdsa==0.19.1
greenlet==3.2.3
h11==0.16.0
idna==3.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
itsdangerous==2.1.2
jinja2>=3.1
python-multipart

